requirements.txt
fastapi
uvicorn
websockets
numpy
pandas
sqlalchemy
sqlmodel
asyncpg
pydantic
pydantic-settings
pyyaml
pytest
httpx